requirements.txt
# Quantitative Trading System Dependencies
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
sqlite3 (stdlib)
python-dateutil>=2.8.0
requests>=2.31.0
websocket-client>=1.6.0
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.0.0
loguru>=0.7.0
